url_scanner_app/requirements.txt
streamlit
pandas
numpy<2.0
tensorflow
xgboost
scikit-learn
tldextract
python-Levenshtein